wwwwww

+++ internal/api/material_type.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException
 
 from internal.adapter.repo.material_type_repo import MaterialTypeRepo
 
@@ -17,13 +17,18 @@ class MaterialType:
         return self.repo.create(material_type, raw_material_loss_percentage)
 
     async def get(self, id):
-        return self.repo.get(id)
+        if self.repo.check(id):
+            return self.repo.get(id)
+        
+        raise HTTPException(status_code=400, detail="does not exist")
 
     async def get_list(self):
         return self.repo.get_list()
 
     async def delete(self, id):
-        return self.repo.delete(id)
+        if self.repo.check(id):
+            return self.repo.delete(id)
+        raise HTTPException(status_code=400, detail="does not exist")
 
     def get_router(self):
         return self.router
