+++ internal/api/products.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException
 
 from internal.adapter.repo.products_repo import ProductsRepo
 
@@ -13,17 +13,22 @@ class Products:
         self.router.get("/get/list", status_code=200)(self.get_list)
         self.router.delete("/delete", status_code=200)(self.delete)
 
-    async def create(self, product_type: str, product_name: str, part_number: str, minimum_partner_cost: str, primary_material: str):
-        return self.repo.create(product_type, product_name, part_number, minimum_partner_cost, primary_material)
+    async def create(self, material_type: str, raw_material_loss_percentage: str):
+        return self.repo.create(material_type, raw_material_loss_percentage)
 
     async def get(self, id):
-        return self.repo.get(id)
+        if self.repo.check(id):
+            return self.repo.get(id)
+        
+        raise HTTPException(status_code=400, detail="does not exist")
 
     async def get_list(self):
         return self.repo.get_list()
 
     async def delete(self, id):
-        return self.repo.delete(id)
+        if self.repo.check(id):
+            return self.repo.delete(id)
+        raise HTTPException(status_code=400, detail="does not exist")
 
     def get_router(self):
         return self.router
